improved visualizer, updated ssl cert, server bug
Node grouping still needs some work.  The application needs to be able
to pull nodes out of grouped folders if a node is being refered from
another edge.  The same is to be said that the grouped folder needs to
skip if a node is already out on the graph.